refactor: 패키지 역할에 맞게 다시 이동
